race format test, class format tweaks
testing races with the top / bottom half page format. not sure it works
yet, things need to be pushed around.

also turned off the class preamble because we weren't using the
alternating colors elsewhere, and the lack of margin flush and indents
was annoying me. stopped new classes on new pages to see how it looked.
also fixed a lot of duplicate colons.